move neovim configuration to init.lua
