+++ modules/tests/neovim.nix
@@ -64,7 +64,7 @@ in
       machine.succeed(
       su('${user}', 'cat ${checkfile}')
       )
-      machine.succeed("cat ${home}/.config/nvim/init.vim")
+      machine.succeed("cat ${home}/.config/nvim/init.lua")
       machine.succeed("cat ${home}/.config/nvim/lua/init.lua")
     '';
   }
